rsu/requirements.txt
pymongo==3.9.0
paho-mqtt==1.5.0

# Worker general
blosc
numpy==1.18.1
scipy
spectrum
sklearn
psutil
joblib
pandas==1.0.1

# Routing
shapely
folium
fastparquet
geopandas
geohash_hilbert

# key-value store